@@ -1,24 +1,23 @@
 name: tool_base
 description: Base for developing command line apps. Take from Flutter Tools.
-version: 1.9.5+3
+version: 2.0.0
 homepage: https://github.com/mmcc007/tool_base
-author: Maurice McCabe <mmcc007@gmail.com>
 
 environment:
-  # The pub client defaults to an <2.0.0 sdk constraint which we need to explicitly overwrite.
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  # To update these, use "flutter update-packages --force-upgrade".
-  archive: ^2.0.9
-  file: ^5.0.8
-  platform: ^2.2.0
-  process: ^3.0.9
-  pub_cache: ^0.2.3
+  archive: ^3.1.6
+  file: ^6.1.2
+  intl: ^0.17.0
+#  platform: ^2.2.0
+  process: ^4.2.4
+  pub_cache: ^0.3.1
 
 dev_dependencies:
-  pedantic: ^1.0.0
-  test: ^1.0.0
-  mockito: ^4.1.0
-  quiver: ^2.0.3
-  file_testing: 2.1.0
+  build_runner: ^2.0.0
+  file_testing: ^3.0.0
+  mockito: ^5.0.16
+  lints: ^1.0.1
+  quiver: ^3.0.0
+  test: ^1.19.2
