@@ -13,11 +13,7 @@ import 'utils.dart';
 
 final AnsiTerminal _kAnsiTerminal = AnsiTerminal();
 
-AnsiTerminal get terminal {
-  return (context == null || context.get<AnsiTerminal>() == null)
-      ? _kAnsiTerminal
-      : context.get<AnsiTerminal>();
-}
+AnsiTerminal get terminal => context.get<AnsiTerminal>() ?? _kAnsiTerminal;
 
 enum TerminalColor {
   red,
@@ -31,20 +27,18 @@ enum TerminalColor {
 
 final OutputPreferences _kOutputPreferences = OutputPreferences();
 
-OutputPreferences get outputPreferences => (context == null || context.get<OutputPreferences>() == null)
-    ? _kOutputPreferences
-    : context.get<OutputPreferences>();
+OutputPreferences get outputPreferences => context.get<OutputPreferences>() ?? _kOutputPreferences;
 
 /// A class that contains the context settings for command text output to the
 /// console.
 class OutputPreferences {
   OutputPreferences({
-    bool wrapText,
-    int wrapColumn,
-    bool showColor,
-  }) : wrapText = wrapText ?? io.stdio?.hasTerminal ?? const io.Stdio().hasTerminal,
+    bool? wrapText,
+    int? wrapColumn,
+    bool? showColor,
+  }) : wrapText = wrapText ?? io.stdio.hasTerminal,
        _overrideWrapColumn = wrapColumn,
-       showColor = showColor ?? platform.stdoutSupportsAnsi ?? false;
+       showColor = showColor ?? platform.stdoutSupportsAnsi;
 
   /// If [wrapText] is true, then any text sent to the context's [Logger]
   /// instance (e.g. from the [printError] or [printStatus] functions) will be
@@ -63,9 +57,9 @@ class OutputPreferences {
   /// and if that's not set, it tries creating a new [io.Stdio] and asks it, if
   /// that doesn't have an idea of the terminal width, then we just use a
   /// default of 100. It will be ignored if [wrapText] is false.
-  final int _overrideWrapColumn;
+  final int? _overrideWrapColumn;
   int get wrapColumn {
-    return  _overrideWrapColumn ?? io.stdio?.terminalColumns
+    return  _overrideWrapColumn ?? io.stdio.terminalColumns
       ?? const io.Stdio().terminalColumns ?? kDefaultTerminalColumns;
   }
 
@@ -105,9 +99,9 @@ class AnsiTerminal {
     TerminalColor.grey: grey,
   };
 
-  static String colorCode(TerminalColor color) => _colorMap[color];
+  static String? colorCode(TerminalColor color) => _colorMap[color];
 
-  bool get supportsColor => platform.stdoutSupportsAnsi ?? false;
+  bool get supportsColor => platform.stdoutSupportsAnsi;
   final RegExp _boldControls = RegExp('(${RegExp.escape(resetBold)}|${RegExp.escape(bold)})');
 
   String bolden(String message) {
@@ -134,7 +128,7 @@ class AnsiTerminal {
     if (!supportsColor || color == null || message.isEmpty)
       return message;
     final StringBuffer buffer = StringBuffer();
-    final String colorCodes = _colorMap[color];
+    final colorCodes = _colorMap[color];
     for (String line in message.split('\n')) {
       // If there were resets in the string before, then keep them, but
       // restart the color right after. This prevents embedded resets from
@@ -165,14 +159,20 @@ class AnsiTerminal {
     }
   }
 
-  Stream<String> _broadcastStdInString;
+  Stream<String>? _broadcastStdInString;
 
   /// Return keystrokes from the console.
   ///
   /// Useful when the console is in [singleCharMode].
   Stream<String> get keystrokes {
-    _broadcastStdInString ??= io.stdin.transform<String>(const AsciiDecoder(allowInvalid: true)).asBroadcastStream();
-    return _broadcastStdInString;
+    var tmp = _broadcastStdInString;
+    if (tmp == null) {
+      tmp = io.stdin
+          .transform<String>(const AsciiDecoder(allowInvalid: true))
+          .asBroadcastStream();
+      _broadcastStdInString = tmp;
+    }
+    return tmp;
   }
 
   /// Prompts the user to input a character within a given list. Re-prompts if
@@ -188,8 +188,8 @@ class AnsiTerminal {
   /// `defaultChoiceIndex`.
   Future<String> promptForCharInput(
     List<String> acceptedCharacters, {
-    String prompt,
-    int defaultChoiceIndex,
+    String? prompt,
+    int? defaultChoiceIndex,
     bool displayAcceptedCharacters = true,
   }) async {
     assert(acceptedCharacters != null);
@@ -203,7 +203,7 @@ class AnsiTerminal {
       charactersToDisplay[defaultChoiceIndex] = bolden(charactersToDisplay[defaultChoiceIndex]);
       acceptedCharacters.add('\n');
     }
-    String choice;
+    String? choice;
     singleCharMode = true;
     while (choice == null || choice.length > 1 || !acceptedCharacters.contains(choice)) {
       if (prompt != null) {
