@@ -8,12 +8,12 @@ import 'package:tool_base/src/base/file_system.dart';
 import '../src/common.dart';
 
 void main() {
-  Config config;
-  Directory tempDir;
+  Config? config;
+  Directory? tempDir;
 
   setUp(() {
     tempDir = fs.systemTempDirectory.createTempSync('flutter_config_test.');
-    final File file = fs.file(fs.path.join(tempDir.path, '.settings'));
+    final File file = fs.file(fs.path.join(tempDir!.path, '.settings'));
     config = Config(file);
   });
 
@@ -23,33 +23,33 @@ void main() {
 
   group('config', () {
     test('get set value', () async {
-      expect(config.getValue('foo'), null);
-      config.setValue('foo', 'bar');
-      expect(config.getValue('foo'), 'bar');
-      expect(config.keys, contains('foo'));
+      expect(config!.getValue('foo'), null);
+      config!.setValue('foo', 'bar');
+      expect(config!.getValue('foo'), 'bar');
+      expect(config!.keys, contains('foo'));
     });
 
     test('get set bool value', () async {
-      expect(config.getValue('foo'), null);
-      config.setValue('foo', true);
-      expect(config.getValue('foo'), true);
-      expect(config.keys, contains('foo'));
+      expect(config!.getValue('foo'), null);
+      config!.setValue('foo', true);
+      expect(config!.getValue('foo'), true);
+      expect(config!.keys, contains('foo'));
     });
 
     test('containsKey', () async {
-      expect(config.containsKey('foo'), false);
-      config.setValue('foo', 'bar');
-      expect(config.containsKey('foo'), true);
+      expect(config!.containsKey('foo'), false);
+      config!.setValue('foo', 'bar');
+      expect(config!.containsKey('foo'), true);
     });
 
     test('removeValue', () async {
-      expect(config.getValue('foo'), null);
-      config.setValue('foo', 'bar');
-      expect(config.getValue('foo'), 'bar');
-      expect(config.keys, contains('foo'));
-      config.removeValue('foo');
-      expect(config.getValue('foo'), null);
-      expect(config.keys, isNot(contains('foo')));
+      expect(config!.getValue('foo'), null);
+      config!.setValue('foo', 'bar');
+      expect(config!.getValue('foo'), 'bar');
+      expect(config!.keys, contains('foo'));
+      config!.removeValue('foo');
+      expect(config!.getValue('foo'), null);
+      expect(config!.keys, isNot(contains('foo')));
     });
   });
 }
