@@ -5,19 +5,19 @@
 import 'dart:async';
 import 'dart:typed_data';
 
+import 'package:quiver/testing/async.dart';
 import 'package:tool_base/src/base/io.dart' as io;
 import 'package:tool_base/src/base/net.dart';
 import 'package:tool_base/src/base/platform.dart';
-import 'package:quiver/testing/async.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
 
 void main() {
   testUsingContext('retry from 500', () async {
-    String error;
+    String? error;
     FakeAsync().run((FakeAsync time) {
-      fetchUrl(Uri.parse('http://example.invalid/')).then((List<int> value) {
+      fetchUrl(Uri.parse('http://example.invalid/')).then((value) {
         error = 'test completed unexpectedly';
       }, onError: (dynamic exception) {
         error = 'test failed unexpectedly: $exception';
@@ -38,9 +38,9 @@ void main() {
   });
 
   testUsingContext('retry from network error', () async {
-    String error;
+    String? error;
     FakeAsync().run((FakeAsync time) {
-      fetchUrl(Uri.parse('http://example.invalid/')).then((List<int> value) {
+      fetchUrl(Uri.parse('http://example.invalid/')).then((value) {
         error = 'test completed unexpectedly';
       }, onError: (dynamic exception) {
         error = 'test failed unexpectedly: $exception';
@@ -61,9 +61,9 @@ void main() {
   });
 
   testUsingContext('retry from SocketException', () async {
-    String error;
+    String? error;
     FakeAsync().run((FakeAsync time) {
-      fetchUrl(Uri.parse('http://example.invalid/')).then((List<int> value) {
+      fetchUrl(Uri.parse('http://example.invalid/')).then((value) {
         error = 'test completed unexpectedly';
       }, onError: (dynamic exception) {
         error = 'test failed unexpectedly: $exception';
@@ -87,9 +87,9 @@ void main() {
   });
 
   testUsingContext('no retry from HandshakeException', () async {
-    String error;
+    String? error;
     FakeAsync().run((FakeAsync time) {
-      fetchUrl(Uri.parse('http://example.invalid/')).then((List<int> value) {
+      fetchUrl(Uri.parse('http://example.invalid/')).then((value) {
         error = 'test completed unexpectedly';
       }, onError: (dynamic exception) {
         error = 'test failed: $exception';
@@ -107,9 +107,9 @@ void main() {
   });
 
   testUsingContext('check for bad override on ArgumentError', () async {
-    String error;
+    String? error;
     FakeAsync().run((FakeAsync time) {
-      fetchUrl(Uri.parse('example.invalid/')).then((List<int> value) {
+      fetchUrl(Uri.parse('example.invalid/')).then((value) {
         error = 'test completed unexpectedly';
       }, onError: (dynamic exception) {
         error = 'test failed: $exception';
@@ -126,15 +126,15 @@ void main() {
       ArgumentError('test exception handling'),
     ),
     Platform: () => FakePlatform.fromPlatform(const LocalPlatform())
-      ..environment = <String, String>{
-        'FLUTTER_STORAGE_BASE_URL': 'example.invalid'
-      },
+      .copyWith(environment: {
+      'FLUTTER_STORAGE_BASE_URL': 'example.invalid'
+    }),
   });
 
   testUsingContext('retry from HttpException', () async {
-    String error;
+    String? error;
     FakeAsync().run((FakeAsync time) {
-      fetchUrl(Uri.parse('http://example.invalid/')).then((List<int> value) {
+      fetchUrl(Uri.parse('http://example.invalid/')).then((value) {
         error = 'test completed unexpectedly';
       }, onError: (dynamic exception) {
         error = 'test failed unexpectedly: $exception';
@@ -158,10 +158,10 @@ void main() {
   });
 
   testUsingContext('max attempts', () async {
-    String error;
-    List<int> actualResult;
+    String? error;
+    List<int>? actualResult;
     FakeAsync().run((FakeAsync time) {
-      fetchUrl(Uri.parse('http://example.invalid/'), maxAttempts: 3).then((List<int> value) {
+      fetchUrl(Uri.parse('http://example.invalid/'), maxAttempts: 3).then((value) {
         actualResult = value;
       }, onError: (dynamic exception) {
         error = 'test failed unexpectedly: $exception';
@@ -269,14 +269,12 @@ class MockHttpClientResponse implements io.HttpClientResponse {
   String get reasonPhrase => '<reason phrase>';
 
   @override
-  StreamSubscription<Uint8List> listen(
-    void onData(Uint8List event), {
-    Function onError,
-    void onDone(),
-    bool cancelOnError,
-  }) {
-    return Stream<Uint8List>.fromFuture(Future<Uint8List>.error(const io.SocketException('test')))
-      .listen(onData, onError: onError, onDone: onDone, cancelOnError: cancelOnError);
+  StreamSubscription<Uint8List> listen(void onData(Uint8List event)?,
+      {Function? onError, void onDone()?, bool? cancelOnError}) {
+    return Stream<Uint8List>.fromFuture(
+            Future<Uint8List>.error(const io.SocketException('test')))
+        .listen(onData,
+            onError: onError, onDone: onDone, cancelOnError: cancelOnError);
   }
 
   @override
