@@ -11,7 +11,7 @@ import '../src/context.dart';
 
 void main() {
   group('OperatingSystemUtils', () {
-    Directory tempDir;
+    Directory? tempDir;
 
     setUp(() {
       tempDir =
@@ -23,7 +23,7 @@ void main() {
     });
 
     testUsingContext('makeExecutable', () async {
-      final File file = fs.file(fs.path.join(tempDir.path, 'foo.script'));
+      final File file = fs.file(fs.path.join(tempDir!.path, 'foo.script'));
       file.writeAsStringSync('hello world');
       os.makeExecutable(file);
 
