@@ -8,15 +8,14 @@ import 'file_system.dart';
 import 'platform.dart';
 
 class Config {
-  Config([File configFile]) {
-    _configFile = configFile ?? fs.file(fs.path.join(_userHomeDir(), '.flutter_settings'));
+  Config([File? configFile]) : _configFile = configFile ?? fs.file(fs.path.join(_userHomeDir(), '.flutter_settings')) {
     if (_configFile.existsSync())
       _values = json.decode(_configFile.readAsStringSync());
   }
 
-  static Config get instance => context.get<Config>();
+  static Config? get instance => context.get<Config>();
 
-  File _configFile;
+  final File _configFile;
   String get configPath => _configFile.path;
 
   Map<String, dynamic> _values = <String, dynamic>{};
