@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:tool_base/src/base/io.dart';
-import 'package:tool_base/src/base/file_system.dart';
-import 'package:tool_base/src/base/os.dart';
 import 'package:mockito/mockito.dart';
-import 'package:process/process.dart';
 import 'package:platform/platform.dart';
+import 'package:process/process.dart';
+import 'package:tool_base/src/base/file_system.dart';
+import 'package:tool_base/src/base/io.dart';
+import 'package:tool_base/src/base/os.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
@@ -17,7 +17,7 @@ const String kPath1 = '/bar/bin/$kExecutable';
 const String kPath2 = '/another/bin/$kExecutable';
 
 void main() {
-  ProcessManager mockProcessManager;
+  ProcessManager mockProcessManager = MockProcessManager();
 
   setUp(() {
     mockProcessManager = MockProcessManager();
@@ -38,7 +38,7 @@ void main() {
       when(mockProcessManager.runSync(<String>['which', 'foo']))
           .thenReturn(ProcessResult(0, 0, kPath1, null));
       final OperatingSystemUtils utils = OperatingSystemUtils();
-      expect(utils.which(kExecutable).path, kPath1);
+      expect(utils.which(kExecutable)?.path, kPath1);
     }, overrides: <Type, Generator>{
       ProcessManager: () => mockProcessManager,
       Platform: () => FakePlatform(operatingSystem: 'linux'),
@@ -73,7 +73,7 @@ void main() {
       when(mockProcessManager.runSync(<String>['where', 'foo']))
           .thenReturn(ProcessResult(0, 0, '$kPath1\n$kPath2', null));
       final OperatingSystemUtils utils = OperatingSystemUtils();
-      expect(utils.which(kExecutable).path, kPath1);
+      expect(utils.which(kExecutable)?.path, kPath1);
     }, overrides: <Type, Generator>{
       ProcessManager: () => mockProcessManager,
       Platform: () => FakePlatform(operatingSystem: 'windows'),
