@@ -26,52 +26,6 @@ export 'package:test/test.dart';
 // TODO(ianh): Remove this once https://github.com/dart-lang/matcher/issues/98 is fixed
 Matcher isInstanceOf<T>() => test_package.TypeMatcher<T>();
 
-void tryToDelete(Directory directory) {
-  // This should not be necessary, but it turns out that
-  // on Windows it's common for deletions to fail due to
-  // bogus (we think) "access denied" errors.
-  try {
-    directory.deleteSync(recursive: true);
-  } on FileSystemException catch (error) {
-    print('Failed to delete ${directory.path}: $error');
-  }
-}
-
-/// Gets the path to the root of the Flutter repository.
-///
-/// This will first look for a `FLUTTER_ROOT` environment variable. If the
-/// environment variable is set, it will be returned. Otherwise, this will
-/// deduce the path from `platform.script`.
-String getFlutterRoot() {
-  if (platform.environment.containsKey('FLUTTER_ROOT'))
-    return platform.environment['FLUTTER_ROOT'];
-
-  Error invalidScript() => StateError('Invalid script: ${platform.script}');
-
-  Uri scriptUri;
-  switch (platform.script.scheme) {
-    case 'file':
-      scriptUri = platform.script;
-      break;
-    case 'data':
-      final RegExp flutterTools = RegExp(r'(file://[^"]*[/\\]flutter_tools[/\\][^"]+\.dart)', multiLine: true);
-      final Match match = flutterTools.firstMatch(Uri.decodeFull(platform.script.path));
-      if (match == null)
-        throw invalidScript();
-      scriptUri = Uri.parse(match.group(1));
-      break;
-    default:
-      throw invalidScript();
-  }
-
-  final List<String> parts = fs.path.split(fs.path.fromUri(scriptUri));
-  final int toolsIndex = parts.indexOf('flutter_tools');
-  if (toolsIndex == -1)
-    throw invalidScript();
-  final String toolsPath = fs.path.joinAll(parts.sublist(0, toolsIndex + 1));
-  return fs.path.normalize(fs.path.join(toolsPath, '..', '..'));
-}
-
 //CommandRunner<void> createTestCommandRunner([FlutterCommand command]) {
 //  final FlutterCommandRunner runner = FlutterCommandRunner();
 //  if (command != null) runner.addCommand(command);
@@ -89,7 +43,7 @@ void updateFileModificationTime(
 }
 
 /// Matcher for functions that throw [ToolExit].
-Matcher throwsToolExit({ int exitCode, Pattern message }) {
+Matcher throwsToolExit({ int? exitCode, Pattern? message }) {
   Matcher matcher = isToolExit;
   if (exitCode != null)
     matcher = allOf(matcher, (ToolExit e) => e.exitCode == exitCode);
