@@ -9,7 +9,7 @@ import 'package:tool_base/tool_base.dart';
 
 Future<T> runInContext<T>(
   FutureOr<T> runner(), {
-  Map<Type, Generator> overrides,
+  Map<Type, Generator>? overrides,
 }) async {
   return await context.run<T>(
     name: 'global fallbacks',
