@@ -24,7 +24,7 @@ void main() {
   group('ToolVersion', () {
     const String settingsFileName = 'settings.json';
     const String toolName = 'sylph';
-    MemoryFileSystem fs;
+    var fs = MemoryFileSystem();
 
     setUp(() {
       fs = MemoryFileSystem();
@@ -36,20 +36,21 @@ void main() {
     });
 
     testUsingContext('get version remotely', () async {
-      final File settingsFile = fs.file(fs.path.join(Cache.flutterRoot, settingsFileName));
+      final File settingsFile = fs.file(fs.path.join(Cache.flutterRoot!, settingsFileName));
       final ToolVersion toolVersion = ToolVersion(toolName, settingsFileName);
       final String version = await toolVersion.getLatestVersion(forceRemote: true);
       final String savedVersion = jsonDecode(settingsFile.readAsStringSync())['latestVersion'];
       expect(version, savedVersion);
     }, overrides: <Type, Generator>{
       FileSystem: () => fs,
-      HttpClientFactory: () => () => MockHttpClient(HttpStatus.ok,
-          result: jsonEncode({
-            "scorecard": {
-              "packageName": "sylph",
-              "packageVersion": "0.7.0+2",
-              "updated": "2019-11-21T22:31:32.194619Z",
-              "packageVersionCreated": "2019-11-20T18:17:13.527154Z",
+      HttpClientFactory: () => () => MockHttpClient(
+          HttpStatus.ok,
+          jsonEncode({
+            'scorecard': {
+              'packageName': 'sylph',
+              'packageVersion': '0.7.0+2',
+              'updated': '2019-11-21T22:31:32.194619Z',
+              'packageVersionCreated': '2019-11-20T18:17:13.527154Z',
             }
           })),
     });
@@ -69,7 +70,7 @@ void main() {
       expect(version, savedVersion);
     }, overrides: <Type, Generator>{
       FileSystem: () => fs,
-      HttpClientFactory: () => () => MockHttpClient(HttpStatus.badRequest, result: jsonEncode(null)),
+      HttpClientFactory: () => () => MockHttpClient(HttpStatus.badRequest, jsonEncode(null)),
     });
 
     test('get installed version', () {
@@ -81,14 +82,14 @@ void main() {
 }
 
 class MockHttpClient implements HttpClient {
-  MockHttpClient(this.statusCode, {this.result});
+  MockHttpClient(this.statusCode, this.result);
 
   final int statusCode;
   final String result;
 
   @override
   Future<HttpClientRequest> getUrl(Uri url) async {
-    return MockHttpClientRequest(statusCode, result: result);
+    return MockHttpClientRequest(statusCode, result);
   }
 
   @override
@@ -98,14 +99,14 @@ class MockHttpClient implements HttpClient {
 }
 
 class MockHttpClientRequest implements HttpClientRequest {
-  MockHttpClientRequest(this.statusCode, {this.result});
+  MockHttpClientRequest(this.statusCode, this.result);
 
   final int statusCode;
   final String result;
 
   @override
   Future<HttpClientResponse> close() async {
-    return MockHttpClientResponse(statusCode, result: result);
+    return MockHttpClientResponse(statusCode, result);
   }
 
   @override
@@ -115,7 +116,7 @@ class MockHttpClientRequest implements HttpClientRequest {
 }
 
 class MockHttpClientResponse implements HttpClientResponse {
-  MockHttpClientResponse(this.statusCode, {this.result});
+  MockHttpClientResponse(this.statusCode, this.result);
 
   @override
   final int statusCode;
@@ -132,10 +133,10 @@ class MockHttpClientResponse implements HttpClientResponse {
 
   @override
   StreamSubscription<Uint8List> listen(
-    void onData(Uint8List event), {
-    Function onError,
-    void onDone(),
-    bool cancelOnError,
+    void onData(Uint8List event)?, {
+    Function? onError,
+    void onDone()?,
+    bool? cancelOnError,
   }) {
     return Stream<Uint8List>.fromIterable(
             <Uint8List>[Uint8List.fromList(result.codeUnits)])
